+++ extensions/expo/src/components/AccountPicker.tsx
@@ -1,31 +1,34 @@
 import { List } from "@raycast/api";
-import { useState, useEffect } from "react";
+import { useEffect } from "react";
 import { AccountsItem } from "../lib/types/users-types.types";
-import { getAccounts } from "../lib/utils";
+import useAuth from "../hooks/useAuth";
 
 export default function AccountPicker({ onPick }: { onPick: (account: AccountsItem) => void }) {
-  const [accounts, setAccounts] = useState<AccountsItem[]>([]);
+  const { accounts } = useAuth();
 
   function onPickAccount(accId: string) {
-    const acc = accounts.find((acc) => acc.id === accId) as AccountsItem;
+    if (!accounts) return;
+
+    const acc = accounts.find((acc) => acc.id === accId);
+    if (!acc) return;
     onPick(acc);
   }
 
   useEffect(() => {
-    async function init() {
-      const userAccounts = await getAccounts();
-      console.log({ userAccounts });
-      setAccounts(userAccounts);
-      onPick(userAccounts[0]);
+    if (accounts && accounts?.length > 0) {
+      onPick(accounts[0]);
     }
-    init();
-  }, []);
+  }, [accounts]);
   return (
     <List.Dropdown tooltip="Account" storeValue onChange={onPickAccount}>
       <List.Dropdown.Section title="Account">
-        {accounts.map((acc) => (
-          <List.Dropdown.Item key={acc.id} title={acc.name} value={acc.id} />
-        ))}
+        {accounts && (
+          <>
+            {accounts.map((acc) => (
+              <List.Dropdown.Item key={acc.id} title={acc.name} value={acc.id} />
+            ))}
+          </>
+        )}
       </List.Dropdown.Section>
     </List.Dropdown>
   );