+++ extensions/things/src/show-today-in-menu-bar.tsx
@@ -21,7 +21,8 @@ export default function ShowTodayInMenuBar() {
   const { data: todos, isLoading, mutate } = useCachedPromise(() => getListTodos('today'));
   const { data: lists } = useCachedPromise(() => getLists());
 
-  const tooltip = todos && todos.length > 0 ? todos[0].name : '';
+  const firstIncompleteTodo = todos?.find((item) => item.status === 'open');
+  const tooltip = firstIncompleteTodo?.name || '';
 
   let title = '';
   if (displayTodo) {
@@ -63,8 +64,12 @@ export default function ShowTodayInMenuBar() {
     <MenuBarExtra icon="things-flat.png" title={title} tooltip={tooltip} isLoading={isLoading}>
       {todos && todos.length > 0 ? (
         <>
-          {displayTodo ? (
-            <MenuBarExtra.Item title="Complete" icon={Icon.CheckCircle} onAction={() => completeTodo(todos[0])} />
+          {displayTodo && firstIncompleteTodo ? (
+            <MenuBarExtra.Item
+              title="Complete"
+              icon={Icon.CheckCircle}
+              onAction={() => completeTodo(firstIncompleteTodo)}
+            />
           ) : null}
           <MenuBarExtra.Section>
             <MenuBarExtra.Item title="Today" />