+++ extensions/awork/CHANGELOG.md
@@ -4,3 +4,7 @@
 
 We're excited to release the first version of our Raycast extension for awork, the powerful, but intuitive project management software from Germany.
 – Hypercode
+
+## [Bug Fix] - 2025-04-15
+
+- Fix error when searching in projects and tasks